Sixth Sprint

+++ search-server/request_queue.h
@@ -35,6 +35,6 @@ private:
 template <typename DocumentPredicate>
 std::vector<Document> RequestQueue::AddFindRequest(const std::string& raw_query, DocumentPredicate document_predicate) {
     const auto result = server_.FindTopDocuments(raw_query, document_predicate);
-    AddRequest(result.size());
+    AddRequest(static_cast<int>(result.size()));
     return result;
 }
